Updates to alerts
